requirements.txt
alabaster==0.7.12
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
blessings==1.7
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
Click==7.0
codecov==2.0.15
colorama==0.4.3
coverage==4.5.4
cycler==0.10.0
docutils==0.16
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
Flask-Cors==3.0.8
furl==2.1.0
future==0.18.2
idna==2.8
imagesize==1.2.0
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
mirakuru==2.2.0
more-itertools==7.2.0
numpy==1.18.1
opencv-python==4.2.0.34
orderedmultidict==1.0.1
packaging==19.2
pamqp==2.3.0
parsedatetime==2.4
pika==1.1.0
Pillow==7.1.0
pluggy==0.13.0
port-for==0.4
prettytable==0.7.2
prettytable-extras==0.1.0
protobuf==3.11.3
psutil==5.7.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.2
pytest==5.3.1
pytest-cov==2.8.1
pytest-rabbitmq==2.0.1
python-dateutil==2.8.1
pytz==2019.3
rabbitmq==0.2.0
rabbitpy==2.0.1
requests==2.22.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==3.0.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stringcase==1.2.0
termgraph==0.2.1
urllib3==1.25.6
virtualenv==16.7.7
wait-for==1.1.1
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
